% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V.pyramid3d.R
\name{V.pyramid3d}
\alias{V.pyramid3d}
\title{3d trophic level volume pyramid}
\usage{
V.pyramid3d(V, TEgm, col = seq(V), alpha = 0.2, add.scale = TRUE,
  scale.len = 10, shift = c(0, 0, 0))
}
\arguments{
\item{V}{vector of volumes (i.e. throughput) from lowest to highers trophic level}

\item{TEgm}{geometric mean transfer efficiency (TLs 2-4)}

\item{col}{colors of V (vector)}

\item{alpha}{alpha of V (single value (0-1))}

\item{add.scale}{logical add scale}

\item{scale.len}{length of side of scale (cube)}

\item{shift}{shift values (x,y,z)}
}
\value{
rgl output
}
\description{
3d trophic level volume pyramid
}
\examples{

### data from Tam et al (2008) - Northern Humboldt Cureent Ecosystem (1995-1996)

# data
TEgm <- 0.1014
Ts <- c(27165, 22349, 1658, 266.3, 17.45, 0.607, 0.00515, 6e-06)
Bs <- c(62.84, 147.1, 70.05, 20.61, 1.658, 0.084, 0.000657, 1e-06, 0)
res <- V2Pdim(V = Ts, TEgm=TEgm)
pal <- colorRampPalette(c("green", "yellow", "red"), bias=3)

\donttest{
# single plot of biomasses by TL
rgl::open3d()
res <- V.pyramid3d(V=Bs[1:5], TEgm = TEgm, col=pal(5), scale.len = 1)

# single plot of throughput by TL
rgl::open3d()
res <- V.pyramid3d(V=Ts[2:5], TEgm = TEgm, col=pal(4), alpha=0.2, scale.len = 10)

# how to set-up a comparison plot (throughout example)
rgl::open3d()
tmp <- V.pyramid3d(V=Ts[2:5], TEgm = TEgm, col=pal(4), add.scale = FALSE)
V.pyramid3d(V=Ts[2:5], TEgm = TEgm, col=pal(4), shift=c(dist(tmp[[1]]$vb[1,1:2])+5,0,0), scale.len = 10)
}

}

