% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2Pdim.R
\name{V2Pdim}
\alias{V2Pdim}
\title{Trophic level volume to pyramid dimensions}
\usage{
V2Pdim(V, TEgm)
}
\arguments{
\item{V}{vector of volumes (i.e. throughput) from lowest to highers trophic level}

\item{TEgm}{geometric mean transfer efficiency (TLs 2-4)}
}
\value{
list containing base widths (a), heights (h), cumulative height (cum.h), 
a:h ratio (ah_ratio), volumes (V), geometric mean transfer efficiency (TEgm), 
cumulative volume (cum.V)
}
\description{
Trophic level volume to pyramid dimensions
}
\examples{
# From Tam et al (2008) - Northern Humboldt Cureent Ecosystem (1995-1996)
TEgm <- 0.1014
Ts <- c(27165, 22349, 1658, 266.3, 17.45, 0.607, 0.00515, 6e-06)
V2Pdim(V = Ts, TEgm = TEgm)
}

