% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Vbox3d.R
\name{Vbox3d}
\alias{Vbox3d}
\title{3d trophic level volume boxes}
\usage{
Vbox3d(V, TL.len = 10, col = seq(V), col.bases.only = FALSE,
  alpha = 0.2, add.scale = TRUE, scale.len = 1, shift = c(0, 0, 0),
  draw = TRUE)
}
\arguments{
\item{V}{vector of volumes (i.e. biomass, throughput) from lowest to highest trophic level}

\item{TL.len}{height of trophic level}

\item{col}{colors of V (vector)}

\item{col.bases.only}{Logical. Color bases of trophic levels only?}

\item{alpha}{alpha of V (single value (0-1))}

\item{add.scale}{logical (TRUE / FALSE) adds scale (cube) to right of box plot}

\item{scale.len}{length of side of scale cube (units of V)}

\item{shift}{shift location of plot (x,y,z). Useful in comparitive plots.}

\item{draw}{Logical. Produce plot (Default: draw = TRUE)}
}
\value{
rgl output
}
\description{
3d trophic level volume boxes
}
\examples{

### data from Tam et al (2008) - Northern Humboldt Current Ecosystem (1995-1996)

# data
TEgm <- 0.1014
Ts <- c(27165, 22349, 1658, 266.3, 17.45, 0.607, 0.00515, 6e-06)
Bs <- c(62.84, 147.1, 70.05, 20.61, 1.658, 0.084, 0.000657, 1e-06, 0)
res <- V2Pdim(V = Ts, TEgm=TEgm)
pal <- colorRampPalette(c("green", "yellow", "red"), bias=3)

\donttest{
# single plot of biomasses by TL
rgl::open3d()
res <- Vbox3d(V=Bs[1:5], TL.len = 2, col=pal(5), scale.len = 1)

# single plot of throughput by TL
rgl::open3d()
res <- Vbox3d(V=Ts[2:5], TL.len = 10, col=pal(4), alpha=0.2, 
 scale.len = 10)

# color bases only
rgl::open3d()
res <- Vbox3d(V=Ts[2:5], TL.len = 10, col=8, col.bases.only = TRUE,
   alpha=0.5, scale.len = 10)

# how to set-up a comparison plot (biomass example)
rgl::open3d()
tmp <- Vbox3d(V=Bs[1:5], TL.len = 2, alpha=0, add.scale = FALSE)
Vbox3d(V=Bs[1:5], TL.len = 2, col=pal(5), 
 shift=c(dist(tmp[[1]]$vb[1,1:2])+5,0,0), scale.len = 1)

}

}
