% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/V2Pdim.R
\name{V2Pdim}
\alias{V2Pdim}
\title{Trophic level volume to pyramid dimensions}
\usage{
V2Pdim(V, TEgm, method = "proportional")
}
\arguments{
\item{V}{vector of volumes (i.e. throughput) from lowest to highest trophic level (TL)}

\item{TEgm}{geometric mean transfer efficiency (TLs 2-4)}

\item{method}{Method to use for determining base to height ratio 
(i.e. pyramid angle). Choices include "proportional" (default) or 
"log". "proportional" is}
}
\value{
A list containing (from lowest to highest TL): base widths (a), heights (h), 
cumulative height (cum.h), a:h ratio (ah_ratio), top angle(angle), volumes (V), 
geometric mean transfer efficiency (TEgm), cumulative volume (cum.V).
}
\description{
Used internally by \link[EwEvis]{Vpyramid3d}. The angle (i.e. steepness) 
of the pyramid is determined by the mean transfer efficiency (\code{TEgm}). 
In the case of method = "proportional", the top angle is proportional
to the inverse of the TEgm value. 
When method = "log", the ratio between the pyramid base length (a) 
and height (h) (i.e. "ah_ratio") is determined as 
\code{ah_ratio = -log10(TEgm)}. Supsequently the top angle is calculated.
}
\examples{
# From Tam et al (2008) - Northern Humboldt Current Ecosystem (1995-1996)
TEgm <- 0.1014
Ts <- c(27165, 22349, 1658, 266.3, 17.45, 0.607, 0.00515, 6e-06)
V2Pdim(V = Ts, TEgm = TEgm)
}

